+++ yash-semantics/CHANGELOG.md
@@ -34,6 +34,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
       `Error` struct has been extended to produce more annotations for errors
       with `Vacancy` information.
 - External dependency versions:
+    - Rust 1.77.0 → 1.79.0
     - yash-syntax 0.10.0 → 0.11.0
 
 ## [0.3.0] - 2024-07-13